backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
mysql-connector-python==8.3.0
pydantic==2.6.4
pydantic-settings==2.2.1
langgraph==0.0.20
google-generativeai==0.4.1
